Update README.md
Originally this file stated that we would start with a data class but I changed it because a network segment is oblivious to what types of data that are being transmit through it. That being said I started with the UDP segment and just fed text data into the class instantiation and encoded it into bytes. 

Later, a data class will be made and that will handle different types of data (right now we are only using text). This documentation now matches what is shown in segment_generator.py more accurately.

I also removed the directions since the code has now been written.